netblock: Initial commit
This library can be used to block network access for programs that are
dynamically linked to glibc. See README for more info.

Do note that most of this code was generated using GPT, and just tweaked
and formatted a bit, and it might be flawed in various ways. This is
also why there is no license attached, as it's simply hard to license
code that has been generated this way.

Signed-off-by: Philip Eklöf <phi@pxy.se>